Update README
Behavior has diverged significantly since forking Brad's version.  Not much
code remains the same.

I wonder if it's time to re-implement this from scratch.  It would be nice to
be able to ditch the Google license.  And it could be built in a less commit-
centric way -- I can see a tool that does this for arbitrary git objects being
useful.